Add user property to product

+++ src/modules/product/product.controller.ts
@@ -13,9 +13,19 @@ import {
   UpdateProductDto
 } from './product.dto';
 import { AppError } from '../../shared/server/http/app-error.util';
+import { RequestWithUser } from '../../shared/middlewares/auth.middleware';
 
 export async function createProductHandler(req: Request, res: Response) {
-  const newProduct = await createProduct(req.body).save();
+  const requestWithUser = req as RequestWithUser;
+
+  if (!requestWithUser.user) {
+    throw new AppError('User not found', HttpCode.NotFound);
+  }
+
+  const newProduct = await createProduct({
+    ...req.body,
+    user: requestWithUser.user._id
+  }).save();
 
   return res.status(HttpCode.Created).json(newProduct);
 }
@@ -49,6 +59,15 @@ export async function updateProductHandler(
   res: Response,
   next: NextFunction
 ) {
+  const requestWithUser = req as RequestWithUser;
+
+  const product = await getProductById(req.params.id);
+
+  if (product?.user.toString() !== requestWithUser.user?._id) {
+    return next(new AppError('Unauthorized', HttpCode.Unauthorized));
+  }
+
+  // TODO: refactor updateProduct to use updateOne instead of findByIdAndUpdate
   const updatedProduct = await updateProduct(req.params.id, req.body);
 
   if (!updatedProduct) {
@@ -63,6 +82,14 @@ export async function deleteProductHandler(
   res: Response,
   next: NextFunction
 ) {
+  const requestWithUser = req as RequestWithUser;
+
+  const product = await getProductById(req.params.id);
+
+  if (product?.user.toString() !== requestWithUser.user?._id) {
+    return next(new AppError('Unauthorized', HttpCode.Unauthorized));
+  }
+
   const { deletedCount } = await deleteProduct(req.params.id);
 
   if (deletedCount === 0) {
